Moving code from BlazingMemoryResource header files to impl files
