+++ engine/CMakeLists.txt
@@ -252,6 +252,7 @@ set(SRC_FILES ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/BlazingH
               ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/taskflow/kernel_type.cpp
               ${CMAKE_SOURCE_DIR}/src/execution_graph/Context.cpp
               ${CMAKE_SOURCE_DIR}/src/bmr/MemoryMonitor.cpp
+              ${CMAKE_SOURCE_DIR}/src/bmr/BlazingMemoryResource.cpp
               ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/taskflow/graph.cpp
               ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/BatchJoinProcessing.cpp
               ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/BatchOrderByProcessing.cpp
