+++ engine/src/bmr/BlazingMemoryResource.h
@@ -43,153 +43,33 @@ class internal_blazing_device_memory_resource : public rmm::mr::device_memory_re
 public:
     // TODO: use another constructor for memory in bytes
 
-	internal_blazing_device_memory_resource(std::string allocation_mode,
+    internal_blazing_device_memory_resource(std::string allocation_mode,
                                             std::size_t initial_pool_size,
                                             std::size_t maximum_pool_size,
                                             std::string allocator_logging_file = "",
-                                            float custom_threshold = 0.95)
-    {
-		total_memory_size = ral::config::gpuTotalMemory();
-		used_memory = 0;
-        memory_limit = (double)custom_threshold * total_memory_size;
+                                            float custom_threshold = 0.95);
 
-        initial_pool_size = initial_pool_size - initial_pool_size % 256; //initial_pool_size required to be a multiple of 256 bytes 
+    virtual ~internal_blazing_device_memory_resource() = default;
 
-        if (total_memory_size <= initial_pool_size) {
-            throw std::runtime_error("Cannot allocate this Pool memory size on the GPU.");
-        }
+    size_t get_memory_used();
+    size_t get_max_memory_used();
+    size_t get_from_driver_used_memory();
+    size_t get_total_memory();
+    size_t get_memory_limit();
+    std::string get_type();
+    bool supports_streams() const noexcept override;
+    bool supports_get_mem_info() const noexcept override;
+    std::string get_full_memory_summary();
 
-        if (allocation_mode == "cuda_memory_resource"){
-            memory_resource_owner = std::make_shared<rmm::mr::cuda_memory_resource>();
-            memory_resource = memory_resource_owner.get();
-        } else if (allocation_mode == "managed_memory_resource"){
-            memory_resource_owner = std::make_shared<rmm::mr::managed_memory_resource>();
-            memory_resource = memory_resource_owner.get();
-        } else if (allocation_mode == "pool_memory_resource") {
-            if (initial_pool_size == 0){
-                memory_resource_owner = rmm::mr::make_owning_wrapper<rmm::mr::pool_memory_resource>(
-                    std::make_shared<rmm::mr::cuda_memory_resource>());
-            } else if (maximum_pool_size == 0) {
-                memory_resource_owner = rmm::mr::make_owning_wrapper<rmm::mr::pool_memory_resource>(
-                    std::make_shared<rmm::mr::cuda_memory_resource>(), initial_pool_size);
-            } else {
-                memory_resource_owner = rmm::mr::make_owning_wrapper<rmm::mr::pool_memory_resource>(
-                    std::make_shared<rmm::mr::cuda_memory_resource>(), initial_pool_size, maximum_pool_size);
-            }            
-            memory_resource = memory_resource_owner.get();
-        } else if (allocation_mode == "managed_pool_memory_resource") {
-            if (initial_pool_size == 0){
-                memory_resource_owner = rmm::mr::make_owning_wrapper<rmm::mr::pool_memory_resource>(
-                    std::make_shared<rmm::mr::managed_memory_resource>());
-            } else if (maximum_pool_size == 0) {
-                memory_resource_owner = rmm::mr::make_owning_wrapper<rmm::mr::pool_memory_resource>(
-                    std::make_shared<rmm::mr::managed_memory_resource>(), initial_pool_size);
-            } else {
-                memory_resource_owner = rmm::mr::make_owning_wrapper<rmm::mr::pool_memory_resource>(
-                    std::make_shared<rmm::mr::managed_memory_resource>(), initial_pool_size, maximum_pool_size);
-            }            
-            memory_resource = memory_resource_owner.get();
-        } else if (allocation_mode == "arena_memory_resource") {
-            if (initial_pool_size == 0){
-                memory_resource_owner = rmm::mr::make_owning_wrapper<rmm::mr::arena_memory_resource>(
-                    std::make_shared<rmm::mr::cuda_memory_resource>());
-            } else if (maximum_pool_size == 0) {
-                memory_resource_owner = rmm::mr::make_owning_wrapper<rmm::mr::arena_memory_resource>(
-                    std::make_shared<rmm::mr::cuda_memory_resource>(), initial_pool_size);
-            } else {
-                memory_resource_owner = rmm::mr::make_owning_wrapper<rmm::mr::arena_memory_resource>(
-                    std::make_shared<rmm::mr::cuda_memory_resource>(), initial_pool_size, maximum_pool_size);
-            }         
-            
-            memory_resource = memory_resource_owner.get();
-        } else if (allocation_mode == "existing"){
-            memory_resource = rmm::mr::get_current_device_resource();        
-        } else {
-            throw std::runtime_error("ERROR creating internal_blazing_device_memory_resource: allocation_mode not recognized.");
-        }
-        type = allocation_mode;
+private:
+    void* do_allocate(size_t bytes, rmm::cuda_stream_view stream) override;
+    void do_deallocate(void* p, size_t bytes, rmm::cuda_stream_view stream) override;
+    bool do_is_equal(device_memory_resource const& other) const noexcept override;
+    std::pair<size_t, size_t> do_get_mem_info(rmm::cuda_stream_view stream) const override;
 
-        if (allocator_logging_file != ""){
-            logging_adaptor.reset(new rmm::mr::logging_resource_adaptor<rmm::mr::device_memory_resource>(
-                memory_resource, allocator_logging_file, /*auto_flush=*/true));
-            memory_resource = logging_adaptor.get();
-        }
-        max_used_memory=0;
-	}
-
-	virtual ~internal_blazing_device_memory_resource() = default;
-
-	size_t get_memory_used() {
-		return used_memory;
-	}
-    size_t get_max_memory_used() {
-        return max_used_memory;
-    }
-    size_t get_from_driver_used_memory() {
-	    return ral::config::gpuUsedMemory();
-    }
-
-	size_t get_total_memory() {
-		return total_memory_size;
-	}
-	size_t get_memory_limit() {
-        return memory_limit;
-    }
-
-    std::string get_type() {
-		return type;
-    }
-
-	bool supports_streams() const noexcept override { return memory_resource->supports_streams(); }
-	bool supports_get_mem_info() const noexcept override { return memory_resource->supports_get_mem_info(); }
-
-    std::string get_full_memory_summary() {
-        std::string summary = "";
-        summary += "Memory Resource Summary:: Type: " + this->type;
-        summary += " | Used Memory: " + std::to_string(this->used_memory);
-        summary += " | Max Used Memory: " + std::to_string(this->max_used_memory);
-        summary += " | Available Memory from driver: " + std::to_string(this->get_from_driver_used_memory());
-        summary += " | Total Memory: " + std::to_string(this->total_memory_size);
-        summary += " | Memory Limit: " + std::to_string(this->memory_limit);
-        return summary;
-    }
-
-private: 
-	void* do_allocate(size_t bytes, rmm::cuda_stream_view stream) override {
-		if (bytes <= 0) { 
-            return nullptr;
-		}
-		used_memory += bytes;
-        if (max_used_memory < used_memory){
-            max_used_memory += bytes;
-        }        
-        
-		return memory_resource->allocate(bytes, stream);
-	}
-
-	void do_deallocate(void* p, size_t bytes, rmm::cuda_stream_view stream) override {
-		if (nullptr == p || bytes == 0) return;
-		if (used_memory < bytes) {
-			std::cerr << "blazing_device_memory_resource: Deallocating more bytes than used right now, used_memory: " << used_memory.load() << " less than " << bytes << " bytes." << std::endl;
-			used_memory = 0;
-		} else {
-			used_memory -= bytes;
-		}
-
-		return memory_resource->deallocate(p, bytes, stream);
-	}
-
-	bool do_is_equal(device_memory_resource const& other) const noexcept override {
-		return memory_resource->is_equal(other);
-	}
-
-	std::pair<size_t, size_t> do_get_mem_info(rmm::cuda_stream_view stream) const override {
-		return memory_resource->get_mem_info(stream);
-	}
-
-	size_t total_memory_size;
-	size_t memory_limit;
-	std::atomic<size_t> used_memory;
+    size_t total_memory_size;
+    size_t memory_limit;
+    std::atomic<size_t> used_memory;
     std::atomic<size_t> max_used_memory;
     std::shared_ptr<rmm::mr::device_memory_resource> memory_resource_owner;
     rmm::mr::device_memory_resource * memory_resource;
@@ -214,38 +94,25 @@ public:
      * 
      * @return blazing_device_memory_resource& the blazing_device_memory_resource singleton
      * ----------------------------------------------------------------------**/
-    static blazing_device_memory_resource& getInstance(){
+    static blazing_device_memory_resource& getInstance() {
         // Myers' singleton. Thread safe and unique. Note: C++11 required.
         static blazing_device_memory_resource instance;
         return instance;
     }
 
-	size_t get_memory_used() {
-		return initialized_resource->get_memory_used();
-	}
+    size_t get_memory_used();
 
-    size_t get_max_memory_used() {
-        return initialized_resource->get_max_memory_used();
-    }
+    size_t get_max_memory_used();
 
-	size_t get_total_memory() {
-		return initialized_resource->get_total_memory() ;
-	}
+    size_t get_total_memory();
 
-    size_t get_from_driver_used_memory()  {
-        return initialized_resource->get_from_driver_used_memory();
-    }
-	size_t get_memory_limit() {
-		return initialized_resource->get_memory_limit() ;
-    }
+    size_t get_from_driver_used_memory();
 
-    std::string get_type() {
-		return initialized_resource->get_type() ;
-    }
+    size_t get_memory_limit();
 
-    std::string get_full_memory_summary() {
-        return initialized_resource->get_full_memory_summary() ;
-    }
+    std::string get_type();
+
+    std::string get_full_memory_summary();
 
   /** -----------------------------------------------------------------------*
    * @brief Initialize RMM options
@@ -268,35 +135,12 @@ public:
                     std::size_t initial_pool_size,
                     std::size_t maximum_pool_size,
                     std::string allocator_logging_file,
-                    float device_mem_resouce_consumption_thresh) {
-        
-        std::lock_guard<std::mutex> guard(manager_mutex);
-
-        // repeat initialization is a no-op
-        if (isInitialized()) return;
-
-        initialized_resource.reset(new internal_blazing_device_memory_resource(
-                allocation_mode, initial_pool_size, maximum_pool_size, 
-                allocator_logging_file, device_mem_resouce_consumption_thresh));
-        
-        rmm::mr::set_current_device_resource(initialized_resource.get());
-        
-        is_initialized = true;
-    }
+                    float device_mem_resouce_consumption_thresh);
 
     /** -----------------------------------------------------------------------*
      * @brief Shut down the blazing_device_memory_resource (clears the context)
      * ----------------------------------------------------------------------**/
-    void finalize(){
-        std::lock_guard<std::mutex> guard(manager_mutex);
-
-        // finalization before initialization is a no-op
-        if (isInitialized()) {
-            registered_streams.clear();
-            initialized_resource.reset();
-            is_initialized = false;
-        }
-    }
+    void finalize();
 
     /** -----------------------------------------------------------------------*
      * @brief Check whether the blazing_device_memory_resource has been initialized.
@@ -304,11 +148,8 @@ public:
      * @return true if blazing_device_memory_resource has been initialized.
      * @return false if blazing_device_memory_resource has not been initialized.
      * ----------------------------------------------------------------------**/
-    bool isInitialized() {
-        return getInstance().is_initialized;
-    }
-   
-   
+    bool isInitialized();
+
 private:
     blazing_device_memory_resource() = default;
     ~blazing_device_memory_resource() = default;
@@ -323,58 +164,31 @@ private:
 };
 
 /**
-	@brief This class represents a custom host memory resource used in the cache system.
+    @brief This class represents a custom host memory resource used in the cache system.
 */
 class internal_blazing_host_memory_resource{
 public:
-	// TODO: percy,cordova. Improve the design of get memory in real time 
-	internal_blazing_host_memory_resource(float custom_threshold) 
-    {
-		struct sysinfo si;
-		if (sysinfo(&si) < 0) {
-            std::cerr << "@@ error sysinfo host "<< std::endl;
-        } 
-        total_memory_size = (size_t)si.freeram;
-        used_memory_size = 0;
-        memory_limit = custom_threshold * total_memory_size;
-	}
+    // TODO: percy,cordova. Improve the design of get memory in real time 
+    internal_blazing_host_memory_resource(float custom_threshold);
 
-	virtual ~internal_blazing_host_memory_resource() = default;
+    virtual ~internal_blazing_host_memory_resource() = default;
 
-    // TODO
-    void allocate(std::size_t bytes)  {
-		used_memory_size +=  bytes;
-	}
+    void allocate(std::size_t bytes);
 
-	void deallocate(std::size_t bytes)  {
-		used_memory_size -= bytes;
-	}
+    void deallocate(std::size_t bytes);
 
-	size_t get_from_driver_used_memory()  {
-        struct sysinfo si;
-		sysinfo (&si);
-        // NOTE: sync point 
-		total_memory_size = (size_t)si.totalram;
-		used_memory_size = total_memory_size - (size_t)si.freeram;;
-        return used_memory_size;
-    }
+    size_t get_from_driver_used_memory();
 
-	size_t get_memory_used() {
-		return used_memory_size;
-	}
+    size_t get_memory_used();
 
-	size_t get_total_memory() {
-		return total_memory_size;
-	}
+    size_t get_total_memory();
 
-    size_t get_memory_limit() {
-        return memory_limit;
-    }
+    size_t get_memory_limit();
 
 private:
     size_t memory_limit;
-	size_t total_memory_size;
-	std::atomic<std::size_t> used_memory_size;
+    size_t total_memory_size;
+    std::atomic<std::size_t> used_memory_size;
 };
 
 /** -------------------------------------------------------------------------*
@@ -390,34 +204,23 @@ public:
      * 
      * @return blazing_host_memory_resource& the blazing_host_memory_resource singleton
      * ----------------------------------------------------------------------**/
-    static blazing_host_memory_resource& getInstance(){
+    static blazing_host_memory_resource& getInstance() {
         // Myers' singleton. Thread safe and unique. Note: C++11 required.
         static blazing_host_memory_resource instance;
         return instance;
     }
 
-	size_t get_memory_used() override {
-		return initialized_resource->get_memory_used();
-	}
+    size_t get_memory_used() override;
 
-	size_t get_total_memory() override {
-		return initialized_resource->get_total_memory() ;
-	}
+    size_t get_total_memory() override;
 
-    size_t get_from_driver_used_memory()  {
-        return initialized_resource->get_from_driver_used_memory();
-    }
-	size_t get_memory_limit() {
-		return initialized_resource->get_memory_limit() ;
-    }
+    size_t get_from_driver_used_memory();
 
-    void allocate(std::size_t bytes)  {
-		initialized_resource->allocate(bytes);
-	}
+    size_t get_memory_limit();
 
-	void deallocate(std::size_t bytes)  {
-		initialized_resource->deallocate(bytes);
-	}
+    void allocate(std::size_t bytes);
+
+    void deallocate(std::size_t bytes);
 
    /** -----------------------------------------------------------------------*
    * @brief Initialize
@@ -428,30 +231,12 @@ public:
    * 
    * @param[in] options Optional options to set
    * ----------------------------------------------------------------------**/
-    void initialize(float host_mem_resouce_consumption_thresh) {
-        
-        std::lock_guard<std::mutex> guard(manager_mutex);
-
-        // repeat initialization is a no-op
-        if (isInitialized()) return;
-
-        initialized_resource.reset(new internal_blazing_host_memory_resource(host_mem_resouce_consumption_thresh));
-
-        is_initialized = true;
-    }
+    void initialize(float host_mem_resouce_consumption_thresh);
 
      /** -----------------------------------------------------------------------*
      * @brief Shut down the blazing_device_memory_resource (clears the context)
      * ----------------------------------------------------------------------**/
-    void finalize(){
-        std::lock_guard<std::mutex> guard(manager_mutex);
-
-        // finalization before initialization is a no-op
-        if (isInitialized()) {
-            initialized_resource.reset();
-            is_initialized = false;
-        }
-    }
+    void finalize();
 
     /** -----------------------------------------------------------------------*
      * @brief Check whether the blazing_device_memory_resource has been initialized.
@@ -459,9 +244,7 @@ public:
      * @return true if blazing_device_memory_resource has been initialized.
      * @return false if blazing_device_memory_resource has not been initialized.
      * ----------------------------------------------------------------------**/
-    bool isInitialized() {
-        return getInstance().is_initialized;
-    }
+    bool isInitialized();
 
 private:
     blazing_host_memory_resource() = default;
@@ -480,48 +263,27 @@ private:
 */
 class blazing_disk_memory_resource : public  BlazingMemoryResource {
 public:
-    static blazing_disk_memory_resource& getInstance(){
+    static blazing_disk_memory_resource& getInstance() {
         // Myers' singleton. Thread safe and unique. Note: C++11 required.
         static blazing_disk_memory_resource instance;
         return instance;
     }
 
-	// TODO: percy, cordova.Improve the design of get memory in real time 
-	blazing_disk_memory_resource(float custom_threshold = 0.75) {
-		struct statvfs stat_disk;
-		int ret = statvfs("/", &stat_disk);
+    // TODO: percy, cordova.Improve the design of get memory in real time 
+    blazing_disk_memory_resource(float custom_threshold = 0.75);
 
-		total_memory_size = (size_t)(stat_disk.f_blocks * stat_disk.f_frsize);
-		size_t available_disk_size = (size_t)(stat_disk.f_bfree * stat_disk.f_frsize);
-		used_memory_size = total_memory_size - available_disk_size;
+    virtual ~blazing_disk_memory_resource() = default;
 
-        memory_limit = custom_threshold *  total_memory_size;
-	}
+    virtual size_t get_from_driver_used_memory();
 
-	virtual ~blazing_disk_memory_resource() = default;
+    size_t get_memory_limit();
 
-	virtual size_t get_from_driver_used_memory()  {
-        struct sysinfo si;
-        sysinfo (&si);
-        // NOTE: sync point 
-        total_memory_size = (size_t)si.totalram;
-        used_memory_size =  total_memory_size - (size_t)si.freeram;
-        return used_memory_size;
-    }
-	size_t get_memory_limit()  {
-        return memory_limit;
-    }
+    size_t get_memory_used();
 
-	size_t get_memory_used() {
-        return used_memory_size;
-	}
-
-	size_t get_total_memory() {
-		return total_memory_size;
-	}
+    size_t get_total_memory();
 
 private:
-	size_t total_memory_size;
+    size_t total_memory_size;
     size_t memory_limit;
-	std::atomic<size_t> used_memory_size;
+    std::atomic<size_t> used_memory_size;
 };
